blog_engine/requirements.txt
Django
Pillow
django-imagekit
transliterate